requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2

# Web scraping
google-play-scraper==1.2.4
requests==2.31.0
beautifulsoup4==4.12.2

# AI/ML
google-generativeai==0.3.2

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
openpyxl==3.1.2

# Development
jupyter==1.0.0
ipykernel==6.27.1
